requirements.txt
pyjwt>=2.10.1
requests>=2.32.4
cryptography>=45.0.5
urllib3>=2.5.0 